src/Python/requirements.txt
mcp>=1.1.2
pymarkdownlnt>=0.9.24
